finish builtin player
